Cleaned and updated
Added the dialog acts and restaurant info in a data folder, and modified the data paths in the .py files so the data files can be found. Removed some unused imports and files.
In the preferences.py file, I added the extract_prefs() method that allows for easy preference extraction in other .py files.
Modified the DialogManagementSystem.py, added new methods. Still a lot of work to do to make it work properly in the big picture